fix: Try something - update Helm chart publishing to use latest version for non-rc tags
From the helm docs seems like they don't intent to support latest tag, which is weird in my opinion, I don't want to keep updating the examples just because of a version bump.

Let's see maybe it works if I just try to rename the package and push it, will I get an error from the registry?

Signed-off-by: Eden Reich <eden.reich@gmail.com>