refactor(mcp): Extract MCP middleware functionality to a dedicated agent (#103)
* refactor(mcp): Extract MCP middleware functionality to a dedicated agent

- It makes sense that the Inference Gateway acts as an agent that connects to MCP and the providers.
- Fix also issues with streaming with MCP now the streaming is passed through to the client correctly while being processed on the Gateway.
- Updated OpenAPI schema to reference the new function call structure.
- Modified `ChatCompletionMessageToolCallChunk` to use the new `ChatCompletionMessageToolCallFunction` type.
- Added comprehensive tests for the agent's functionality, including tool execution and streaming.
- Enhanced MCP middleware tests to improve logging and error handling.
- Generated mock for the Agent interface to facilitate testing.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* docs(mcp): Improve the comment for HTTP responses

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* docs(agent): Remove redundant comments in RunWithStream and ExecuteTools methods

There is a log entry and it's clearly what the code is doing.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* fix(logger): Add caller skip to logger configuration to show where the actual code was called and not the wrapper code

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor(agent): Change log level from Debug to Info for tool execution logging

I think I would like to see it also in production, it's a useful information in order to know the flow of the request lifecycle.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor(tests): Change log level from Debug to Info for tool execution logging in agent tests

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor(streaming): Enhance RunWithStream to handle multiple tool call iterations and send a single AGENT_DONE signal

fix(mcp): Update middleware to filter out [DONE] markers and ensure proper streaming response handling
test(mcp): Add test for multiple tool call iterations to validate AGENT_DONE signal and response content

Pass the stream from provider AS-IS, less parsing, more efficient - if there are special cases I'll test it.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor(agent): Update completion signal to [DONE] and enhance chunk processing logic

Simplify it, I though I would need an extra [AGENT_DONE] but actually I could just filter out the [DONE] from the subsequent requests to provider and write the [DONE] when the agent actually done.

TODO - I still need to refactor the part where I parse the data, it's seems a bit clanky. I tested it with Groq and DeepSeek and I had to add this part when using DeepSeek not entirely sure yet why, but I'll investigate.

Signed-off-by: Eden Reich <eden.reich@gmail.com>

* refactor(tests): Cleanup - remove unnecessary logging in streaming tool call iteration test

Signed-off-by: Eden Reich <eden.reich@gmail.com>

---------

Signed-off-by: Eden Reich <eden.reich@gmail.com>